% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdms.R
\docType{class}
\name{TdmsSegmentObject}
\alias{TdmsSegmentObject}
\title{TdmsSegmentObject class}
\format{An \code{\link{R6Class}} generator object}
\usage{
TdmsSegmentObject
}
\description{
TdmsSegmentObject class
}
\keyword{data}

